+++ src/drawShape.cpp
@@ -1,7 +1,7 @@
 #include "drawShape.hpp"
 
 
-DrawShape::DrawShape(const cv::Point2f& position, Color color, float velocityY)
+DrawShape::DrawShape(const cv::Point2f& position, Color color, int velocityY)
     : Shape(position, color, velocityY)
 {
 }
